add security call

+++ employee-service/src/main/java/com/enterprise/management/service/EmployeeService.java
@@ -13,5 +13,5 @@ public interface EmployeeService {
 
     Employee saveEmployee(Employee employee);
 
-    Employee findEmployeeById(Integer id);
+    Employee findEmployeeById(Long id);
 }
